fixed api request

+++ prepper.sh
@@ -159,10 +159,12 @@ function mashup_prep() {
         rm -f *.zip
     done
 
-    read -p "Set Google api key? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || return
+    if [ $prepped = false ]; then
+        read -p "Set Google api key? (Y/N): " confirm && [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]] || return
 
-    read -p "What is your api key?" api_key
-    echo -e "export API_KEY=$api_key" > google_keys
+        read -p "What is your api key?" api_key
+        echo -e "export API_KEY=$api_key" > google_keys
+    fi
     return
 }
